Massive update
Removed a lot of functions relating to Core.Address from the sandbox.
Address should not be dealt with by the core and I will phase it out
over time. Added the ability to make sure all requests received by
Core.Ajax have a header attached (useful in detecting 300 redirects)
Core.Communication now returns an array of responses from every
notification sent. Core.Databinding (which relies on knockout.js and
therefore should probably be removed) now doesn't attempt to remove
bindings if the element is not found in the DOM. Core.ModuleGrouping now
can raise events on the shutdown of module groups. Core.Storage now
tests if local storage is full (via a try/catch) and clears local
storage if it is. Added a helper class for building URL's. (should be in
another place, need associated utilities)